Fixed issue #10 -renamed heder files -moved from .h to .hpp convention
