+++ main.cpp
@@ -12,7 +12,7 @@
 #include <stdio.h>
 #include <GLFW/glfw3.h> // Will drag system OpenGL headers
 
-#include "GuiPlugApp.h"
+#include "GuiPlugApp.hpp"
 
 static void glfw_error_callback(int error, const char *description)
 {
